+++ app/src/main/java/notify/gotify/ClearNotificationsReceiver.kt
@@ -1,0 +1,12 @@
+package notify.gotify
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import androidx.core.app.NotificationManagerCompat
+
+class ClearNotificationsReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent?) {
+        NotificationManagerCompat.from(context).cancelAll()
+    }
+}